Move test to ComputedFolderTest class.
