+++ src/test/java/com/cloudbees/hudson/plugins/folder/computed/ComputedFolderTest.java
@@ -24,6 +24,7 @@
 package com.cloudbees.hudson.plugins.folder.computed;
 
 import com.cloudbees.hudson.plugins.folder.AbstractFolderDescriptor;
+import com.cloudbees.hudson.plugins.folder.Folder;
 import com.cloudbees.hudson.plugins.folder.views.AbstractFolderViewHolder;
 import com.gargoylesoftware.htmlunit.html.DomElement;
 import com.gargoylesoftware.htmlunit.html.HtmlPage;
@@ -66,9 +67,9 @@ import java.util.Set;
 import java.util.TreeMap;
 import java.util.TreeSet;
 import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
 import javax.servlet.ServletException;
 import org.apache.commons.lang.StringUtils;
+import org.junit.Assert;
 import org.junit.Test;
 
 import static org.hamcrest.Matchers.containsInAnyOrder;
@@ -102,7 +103,11 @@ public class ComputedFolderTest {
         assertEquals("[A, B, C]", d.created.toString());
 
         // ComputedFolder page opens correctly
-        r.createWebClient().getPage(d);
+        try {
+            r.createWebClient().getPage(d);
+        } catch (Exception ex) {
+            Assert.fail("ComputedFolder<FreeStyleProject> cannot be opened: " + ex.getMessage());
+        }
 
         d.recompute(Result.SUCCESS);
         d.assertItemNames(3, "A", "B", "C");
@@ -122,6 +127,29 @@ public class ComputedFolderTest {
             descriptions.put(p.getName(), p.getDescription());
         }
         assertEquals("{A=updated in round #5, B=created in round #5, C=updated in round #5, D=created in round #5}", descriptions.toString());
+    }
+
+    @Test
+    public void foldersAsChildren() throws Exception {
+        final SampleComputedFolderWithFoldersAsChildren d = r.jenkins.createProject(SampleComputedFolderWithFoldersAsChildren.class, "d");
+        d.recompute(Result.SUCCESS);
+        d.kids.addAll(Arrays.asList("A"));
+        d.recompute(Result.SUCCESS);
+
+        // Folder page opens correctly
+        try {
+            Folder a = d.getItems().iterator().next();
+            r.createWebClient().getPage(a);
+        } catch (Exception ex) {
+            Assert.fail("Folder inside ComputedFolder cannot be opened: " + ex.getMessage());
+        }
+
+        // ComputerFolder page does no open
+        try {
+            r.createWebClient().getPage(d);
+        } catch (Exception ex) {
+            Assert.fail("ComputedFolder<Folder> cannot be opened: " + ex.getMessage());
+        }
     }
 
     @Test
@@ -557,6 +585,60 @@ public class ComputedFolderTest {
             @Override
             public TopLevelItem newInstance(ItemGroup parent, String name) {
                 return new SampleComputedFolder(parent, name);
+            }
+
+        }
+
+    }
+
+    @SuppressWarnings({"unchecked", "rawtypes"})
+    public static class SampleComputedFolderWithFoldersAsChildren extends ComputedFolder<Folder> {
+
+        List<String> kids = new ArrayList<String>();
+        int round;
+
+        private SampleComputedFolderWithFoldersAsChildren(ItemGroup parent, String name) {
+            super(parent, name);
+        }
+
+        @Override
+        protected void computeChildren(ChildObserver<Folder> observer, TaskListener listener) throws IOException, InterruptedException {
+            round++;
+            listener.getLogger().println("=== Round #" + round + " ===");
+            for (String kid : kids) {
+                listener.getLogger().println("considering " + kid);
+                Folder p = observer.shouldUpdate(kid);
+                try {
+                    if (p == null) {
+                        if (observer.mayCreate(kid)) {
+                            listener.getLogger().println("creating a child");
+                            p = new Folder(this, kid);
+                            p.setDescription("created in round #" + round);
+                            observer.created(p);
+                        } else {
+                            listener.getLogger().println("not allowed to create a child");
+                        }
+                    } else {
+                        listener.getLogger().println("updated existing child with description " + p.getDescription());
+                        p.setDescription("updated in round #" + round);
+                    }
+                } finally {
+                    observer.completed(kid);
+                }
+            }
+
+        }
+
+        String recompute(Result result) throws Exception {
+            return doRecompute(this, result);
+        }
+
+        @TestExtension
+        public static class DescriptorImpl extends AbstractFolderDescriptor {
+
+            @Override
+            public TopLevelItem newInstance(ItemGroup parent, String name) {
+                return new SampleComputedFolderWithFoldersAsChildren(parent, name);
             }
 
         }
